[Perf][Elementwise] Polynomial erf, one divisor per sinusoidal pair, and a second load for latency-bound unaries (#1997)
H200, CI image, CUPTI device-busy median, `benchmarks/` timing. Ratio is
against the fastest baseline tag on each row.

| op / workload | before | after | ratio |
| --- | --- | --- | --- |
| Sinusoidal 4096x4096 float16 | 81.8us | 25.2us | 2.43 -> 4.43 |
| Erf 256M float16 | 420.8us | 303.9us | 0.92 -> 1.27 |
| Gelu 2048x14336 float16 | 52.8us | 38.7us | 0.91 -> 1.24 |
| GeluAndMul 2048x14336 float16 | 55.0us | 44.5us | 1.09 -> 1.35 |
| Sigmoid 256M float16 | 301.6us | 252.9us | 1.07 -> 1.00* |
| Sin / Cos 256M float16 | 364 / 378us | 322 / 332us | 0.98 -> 1.17 /
1.18 |
| Tanh 256M float16 | 296.0us | 274.0us | 0.99 -> 1.07 |
| Silu 2048x14336 float16 | 31.6us | 30.9us | 0.97 -> 1.02 |
| Mish 16x256x80x80 float16 | 40.2us | 36.1us | 1.58 -> 1.76 |
| Log / Log1p 16M float32 | 35.7 / 35.1us | 34.0 / 33.9us | 0.96 / 0.97
-> 1.00 |
| Hardswish 32x96x56x56 bfloat16 | 13.2us | 11.7us | 0.88 -> 1.01 |
| Hardsigmoid 32x960 float16 | 1.7us | 1.4us | 0.89 -> 1.07 |
| Sign 256M float16 | 263.5us | 251.6us | 0.97 -> 1.00 |
| isnan / isinf / isfinite / logical_not 16M float32 | 25.6us | 23.4us |
0.92 -> 1.00 |

\* sigmoid's baseline on that row is torch-compile, which the earlier
1.07 was not measured against.

On `(16, 256, 56, 56)` against a `(256, 1, 1)` broadcast:

| op | float16 | float32 |
| --- | --- | --- |
| Eq / Ne / Lt / Le / Gt / Ge, LogicalAnd / Or | 17.2 -> 13.2us, 0.69 ->
0.89 | 21.6 -> 18.8us, 0.85 -> 0.97 |
| Minimum / Maximum | 34.2 -> 16.9us, 0.42 -> 0.85 | 0.85 -> 0.94 |
| Div | 17.1 -> 15.8us, 0.83 -> 0.90 | 0.90 -> 0.99 |

Reductions on `(4, 128, 4096)` over axes 0 and 2:

| op | before | after | ratio |
| --- | --- | --- | --- |
| Sum / Mean | 5.9 / 5.8us | 4.7 / 4.8us | 0.81 / 0.82 -> 1.02 / 1.01 |
| Amax / Amin | 5.8us | 4.8us | 0.89 / 0.87 -> 1.09 / 1.06 |
| CountNonzero | 6.4us | 5.1us | 0.74 -> 0.94 |

Over the whole elementwise and reduction benchmark, 456 rows: median
1.006, 333 at or above the fastest tag, 424 within 5% of it.

## changes

- erf for float16/bfloat16 is a degree-10 polynomial, saturating to
exactly +-1 so GELU's tail stays exact; float32 keeps `erff`.
- Hardswish and Hardsigmoid multiply by `1/6`, Div computes in float32,
Sigmoid and Silu use `exp2` — each matching what torch does.
- Sign, Mish and min/max drop a `T.if_then_else`, which scalarises the
element loop.
- Sinusoidal computes `pow` once per dimension pair instead of once per
element.
- A transcendental body carries 32 bytes per thread, keeping a second
load in flight; the width bump a narrow output earns is skipped where it
would strand a row-broadcast tail.
- A broadcast block stages through fragments when its body cannot
vectorise.
- The edge-axis reduce writes its result dtype directly, dropping a
third kernel launch.
- 18 test nodes. `nightly.yml` loses its `schedule:`, leaving
`workflow_dispatch`.

+++ src/tileops/kernels/reduction/logical_reduce.py
@@ -456,7 +456,9 @@ class LogicalReduceKernel(Kernel):
         partials = stage(cfg["block_m"], cfg["threads"])(x.reshape(lead * kept, trail))
         partials = partials.reshape(lead, kept)
         if self.op_kind == "count_nonzero":
-            return reduce_down_rows(partials, "sum", "float32", "float32", 0.0).to(torch.int64)
+            # The columns pass writes int64 itself; leaving it in fp32 and casting
+            # after costs a third kernel launch.
+            return reduce_down_rows(partials, "sum", "float32", "int64", 0.0)
         outer_kind = "amax" if self.op_kind == "any" else "amin"
         return reduce_down_rows(partials, outer_kind, "int8", "int8", 0.0).view(torch.bool)
